backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
pydantic==2.7.0
sqlalchemy==2.0.30
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
email-validator==2.1.1
